Changed point value to double data type for partial point values

+++ src/LinkedQuestion.cpp
@@ -13,7 +13,7 @@ struct LinkedQuestion{
     std::string questionContent;
     QuestionType questionType;
     LinkedQuestion *nextQuestion;
-    int pointValue;
+    double pointValue;
 
     // Not implemented
     struct LinkedAnswer{
